1) Refactoring, split unit tests to make them more manageable. 2) Skeleton code for payload extraction application
